Populate the content with the JSON file
The content of the summary-categories-container was created dynamically
by using the template element to hold a category-component, clone that
template, fetch de data and fill the template content with its
corresponding data and later append it to the
summary-categories-container.

Also the average score was obtained by fetching the data and process
it to obtain the average.